feat(structure): introduce hook directory
