+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod app;
 pub mod components;
 pub mod contexts;
 pub mod functions;
+pub mod hooks;
 pub mod views;
 
 use cfg_if::cfg_if;
